Add support for CPU and heap profiling.
These changes add a new flag to the CLI, `--profiling`, that enables
CPU and heap profiling of the CLI itself. The argument to this flag
serves as a prefix for the CPU and heap profile outputs; these profiles
are written to `[filename].[pid].cpu` and `[filename].[pid].mem`,
respectively.

These changes also fix an issue with `cmdutil.RunFunc` wherein any error
would prevent the command's post-run hooks from executing.