Use select to notice timeouts
Before we were reading a boolean variable that was set by another goroutine.
Without any sync primitives, I think Go is within its rights to optimize
that away.

We now require that `ctx` is non-`nil`. This was already true for all callers
and consistent with other Go functions that take contexts.