Run refresh in integration test framework
Now that refresh is alive for managed stacks, we can do a refresh
in the integration test framework.  We also add the --expect-no-changes
flag to refresh and use it to validate that no changes result from the
refresh.  Since it's possible that some resources will always yield
refresh changes (e.g., perhaps a last updated timestamp), we also
support an ExpectRefreshChanges flag for tests to opt into.