Propagate ID in dynamic provider's ReadResponse
The ReadResponse RPC evolved a bit over time and apparently the dynamic
provider fell behind after the initial cut.  In particular, its lack of
an ID in the ReadResponse tricked the engine into thinking that all
dynamic resources went missing during a refresh.  The fix is simple:
propagate the ID.