Add an experimental run command
This adds an experimental run command, tucked beneath the debug-only
CLI envvar.  The idea here is to experiment with the notion that a
Pulumi program can operate just like a normal program you'd execute on
your machine, except that it executes in the cloud.  To do this, we
simply new up an anonymous stack, configure it, deploy the program, tail
the logs, and then, upon ^C, tear it all down again afterwards.  This
works remarkably well for "simple" stacks -- especially serverless ones
-- but obviously not so muchf or stacks that take eons to stand up.