More progress
Introduce Pulum.Input<T> which is either a T or a Task<T> (in the
future, it could also be an Output<T>).

Code (by hand) something that looks like what we might end up using
for `tfgen'`d projections of Bucket and Bucket content. A bunch of
stuff is missing there, but it is sufficent to change the example code
to look a little more in line with what you'd actually end up writing.

The next thing I plan on doing is moving away from CSI, in favor of an
actual .NET binary, so I have nice IntelliSense in the example.