Rework registration.
Instead of using .ContinueWith all over the place, register a single
ContinueWith call when we begin resource registration. When it
completes, we then call a callback on each resource that uses a
TaskCompletionSource to complete a task that represent the output
property.

This reduces the number of continuations we'll have floating around
and makes the code a little easier to grok.