Remove the need for /opt/pulumi
This is an artifact of the past, where we did not ship packages via
NPM but instead linked everything. At that time, we wanted to have our
own install location for Pulumi, because it reflected how we shipped
the product.

This was pretty confusing, because we had both `build` and `install`
targets, and we'd put some binaries both in `GOBIN` and
`/opt/pulumi/bin` leading to general confusion.

Now we simplify things. We install everything into GOBIN (both our go
binaries and the language executor scripts) and when we `yarn link` to
register packages we just link the version we built into the bin
folder.

Overall, this should simplify our story and make things a little more
regular.

We no longer require any directories to be created before building
Pulumi.