Bump versions in compat test
There is not a prebuilt binary for Node 10.7 for the older version of gRPC
that we had locked to, forcing a rebuild when the project was restored.
On some CI workers this does not work out and the test spews a bunch of
output and then appears to hang.

Since this doesn't seem related to Pulumi in any way, just bump our
version.